Ran tests on different sized images
Created a loop to test out all error redundancies for sizes up to 10.

Still prints results to screen so need to fix that.

Also now creates 8 images (code_1 -> code_4 also with _blurred) as I
feel this is the minimum required to work efficiently.

NOTE: can be very slow to add the blur to larger images.